+++ src/main/java/op/banco/dto/CuentaYanki.java
@@ -1,0 +1,33 @@
+package op.banco.dto;
+
+import lombok.Getter;
+import lombok.Setter;
+import lombok.ToString;
+
+@Getter
+@Setter
+@ToString
+public class CuentaYanki {
+	
+	private String id;
+	private String dni;
+	private String numeroCelular;	
+	private TipoCuentaBanco tipoProducto;
+	private String fecha_afiliacion;
+	private String fecha_caducidad;
+	private double saldo;
+	private String codigoBanco;
+	
+	
+	
+}
+
+
+
+
+
+
+
+
+
+
