+++ src/main/java/op/banco/dto/TipoCuentaBanco.java
@@ -1,0 +1,17 @@
+package op.banco.dto;
+
+
+
+import lombok.Getter;
+import lombok.Setter;
+import lombok.ToString;
+
+@Getter
+@Setter
+@ToString
+public class TipoCuentaBanco {
+	
+	private String id;
+	private String descripcion;
+	
+}
