+++ src/main/java/op/banco/events/Event.java
@@ -1,0 +1,15 @@
+package op.banco.events;
+
+import lombok.Data;
+import lombok.ToString;
+
+import java.util.Date;
+
+@ToString
+@Data
+public abstract class Event <T> {
+    private String id;
+    private Date date;
+    private EventType type;
+    private T data;
+}
