+++ src/main/java/op/banco/events/EventType.java
@@ -1,0 +1,5 @@
+package op.banco.events;
+
+public enum EventType {
+	CREATED, UPDATED, DELETED
+}
