+++ src/main/java/op/banco/events/OperacionCreatedEvent.java
@@ -1,0 +1,10 @@
+package op.banco.events;
+
+import lombok.Data;
+import lombok.EqualsAndHashCode;
+
+@Data
+@EqualsAndHashCode(callSuper = true)
+public class OperacionCreatedEvent extends Event<Operacion> {
+
+}
