+++ src/main/java/op/banco/exception/Exception.java
@@ -1,0 +1,17 @@
+package op.banco.exception;
+
+import java.time.ZonedDateTime;
+
+import org.springframework.http.HttpStatus;
+
+import lombok.*;
+
+@Data
+@AllArgsConstructor
+public class Exception {
+	
+	private final String message;
+	private final HttpStatus httpstatus;
+	private final ZonedDateTime timestamp;
+	
+}
