+++ src/main/java/op/banco/exception/RequestException.java
@@ -1,0 +1,17 @@
+package op.banco.exception;
+
+
+public class RequestException extends RuntimeException {
+
+	public RequestException(String message)
+	{
+	  super(message);
+	}
+	
+	public RequestException(String message, Throwable cause)
+	{
+	  super(message, cause);
+	}
+	
+	
+}
