+++ src/main/java/op/banco/exception/ResponseStatus.java
@@ -1,0 +1,15 @@
+package op.banco.exception;
+
+public class ResponseStatus extends RuntimeException{
+	
+	public ResponseStatus(String message)
+	{
+	  super(message);
+	}
+	
+	public ResponseStatus(String message, Throwable cause)
+	{
+	  super(message, cause);
+	}
+	
+}
