--- src/main/java/spring/boot/webflu/ms/op/banco/app/dao/TipoOperacionBancoDao.java
@@ -1,9 +1,0 @@
-package spring.boot.webflu.ms.op.banco.app.dao;
-
-import org.springframework.data.mongodb.repository.ReactiveMongoRepository;
-
-import spring.boot.webflu.ms.op.banco.app.documents.TipoOperacionBanco;
-
-public interface TipoOperacionBancoDao extends ReactiveMongoRepository<TipoOperacionBanco, String> {
-
-}
