--- src/main/java/spring/boot/webflu/ms/op/banco/app/dto/TipoProducto.java
@@ -1,17 +1,0 @@
-package spring.boot.webflu.ms.op.banco.app.dto;
-
-
-
-import lombok.Getter;
-import lombok.Setter;
-import lombok.ToString;
-
-@Getter
-@Setter
-@ToString
-public class TipoProducto {
-	
-	private String id;
-	private String descripcion;
-	
-}
