--- src/main/java/spring/boot/webflu/ms/op/banco/app/exception/Exception.java
@@ -1,17 +1,0 @@
-package spring.boot.webflu.ms.op.banco.app.exception;
-
-import java.time.ZonedDateTime;
-
-import org.springframework.http.HttpStatus;
-
-import lombok.*;
-
-@Data
-@AllArgsConstructor
-public class Exception {
-	
-	private final String message;
-	private final HttpStatus httpstatus;
-	private final ZonedDateTime timestamp;
-	
-}
