--- src/main/java/spring/boot/webflu/ms/op/banco/app/exception/RequestException.java
@@ -1,17 +1,0 @@
-package spring.boot.webflu.ms.op.banco.app.exception;
-
-
-public class RequestException extends RuntimeException {
-
-	public RequestException(String message)
-	{
-	  super(message);
-	}
-	
-	public RequestException(String message, Throwable cause)
-	{
-	  super(message, cause);
-	}
-	
-	
-}
