+++ src/main/resources/application.properties
@@ -1,13 +1,2 @@
-server.port=8023
 
-spring.data.mongodb.uri=mongodb://localhost:27017/spring_boot_ms_op_bank
-spring.application.name = ms-op-bank
 
-configuracion.texto=Configurando ambiente ambiente por defecto
-
-eureka.client.service-url.defaultZone = http://localhost:8761/eureka
-
-#timeout
-hystrix.command.default.execution.isolation.thread.timeoutInMilliseconds: 20000
-ribbon.ConnectTimeout: 3000
-ribbon.ReadTimeout: 10000
